backend/requirements.txt
flask
flask-cors
boto3
pyjwt
bcrypt
qrcode
pillow
pytest
requests
qr_ticket_manager
